Move AWI-GRPC-Catalyst-SDWAN with operator to kube-awi
This change moves the charts responsible for deploying
AWI-GRPC-Catalyst-SDWAN controller and the k8s operator to the
kube-awi repository.

This repository, embeds the charts from kube-awi creating an
addition to the operator setup with a visibility provided by
awi-infra-guard service, the UI service that uses this visibility
mode and envoy-proxy that proxies requests for UI and backend
services and allows CLI to handle both AWI-GRPC-Catalyst-SDWAN
and awi-infra-guard requests.